Dota2_Ability_Arena_avoider/requirements.txt
requests
pyautogui
pytesseract
Pillow
transliterate